+++ packages/alfa-collection/src/collection.ts
@@ -1,5 +1,6 @@
 import { Applicative } from "@siteimprove/alfa-applicative";
-import { Comparer } from "@siteimprove/alfa-comparable";
+import { Callback } from "@siteimprove/alfa-callback";
+import { Comparable, Comparer, Comparison } from "@siteimprove/alfa-comparable";
 import { Equatable } from "@siteimprove/alfa-equatable";
 import { Foldable } from "@siteimprove/alfa-foldable";
 import { Functor } from "@siteimprove/alfa-functor";
@@ -13,6 +14,8 @@ import { Predicate } from "@siteimprove/alfa-predicate";
 import { Reducer } from "@siteimprove/alfa-reducer";
 import { Refinement } from "@siteimprove/alfa-refinement";
 
+const { compareComparable } = Comparable;
+
 export interface Collection<T>
   extends Functor<T>,
     Monad<T>,
@@ -23,6 +26,7 @@ export interface Collection<T>
     Serializable {
   readonly size: number;
   isEmpty(): this is Collection<never>;
+  forEach(callback: Callback<T>): void;
   map<U>(mapper: Mapper<T, U>): Collection<U>;
   flatMap<U>(mapper: Mapper<T, Collection<U>>): Collection<U>;
   reduce<U>(reducer: Reducer<T, U>, accumulator: U): U;
@@ -37,6 +41,7 @@ export interface Collection<T>
   collect<U>(mapper: Mapper<T, Option<U>>): Collection<U>;
   collectFirst<U>(mapper: Mapper<T, Option<U>>): Option<U>;
   some(predicate: Predicate<T>): boolean;
+  none(predicate: Predicate<T>): boolean;
   every(predicate: Predicate<T>): boolean;
   count(predicate: Predicate<T>): number;
   distinct(): Collection<T>;
@@ -47,6 +52,7 @@ export namespace Collection {
     // Collection<T> methods
 
     isEmpty(): this is Keyed<K, never>;
+    forEach(callback: Callback<V, void, [K]>): void;
     map<U>(mapper: Mapper<V, U, [K]>): Keyed<K, U>;
     flatMap<U>(mapper: Mapper<V, Keyed<K, U>, [K]>): Keyed<K, U>;
     reduce<U>(reducer: Reducer<V, U, [K]>, accumulator: U): U;
@@ -63,6 +69,7 @@ export namespace Collection {
     collect<U>(mapper: Mapper<V, Option<U>, [K]>): Keyed<K, U>;
     collectFirst<U>(mapper: Mapper<V, Option<U>, [K]>): Option<U>;
     some(predicate: Predicate<V, [K]>): boolean;
+    none(predicate: Predicate<V, [K]>): boolean;
     every(predicate: Predicate<V, [K]>): boolean;
     count(predicate: Predicate<V, [K]>): number;
     distinct(): Keyed<K, V>;
@@ -80,6 +87,7 @@ export namespace Collection {
     // Collection<T> methods
 
     isEmpty(): this is Unkeyed<never>;
+    forEach(callback: Callback<T>): void;
     map<U>(mapper: Mapper<T, U>): Unkeyed<U>;
     flatMap<U>(mapper: Mapper<T, Unkeyed<U>>): Unkeyed<U>;
     reduce<U>(reducer: Reducer<T, U>, accumulator: U): U;
@@ -94,6 +102,7 @@ export namespace Collection {
     collect<U>(mapper: Mapper<T, Option<U>>): Unkeyed<U>;
     collectFirst<U>(mapper: Mapper<T, Option<U>>): Option<U>;
     some(predicate: Predicate<T>): boolean;
+    none(predicate: Predicate<T>): boolean;
     every(predicate: Predicate<T>): boolean;
     count(predicate: Predicate<T>): number;
     distinct(): Unkeyed<T>;
@@ -111,6 +120,7 @@ export namespace Collection {
     // Collection<T> methods
 
     isEmpty(): this is Indexed<never>;
+    forEach(callback: Callback<T, void, [number]>): void;
     map<U>(mapper: Mapper<T, U, [number]>): Indexed<U>;
     flatMap<U>(mapper: Mapper<T, Indexed<U>, [number]>): Indexed<U>;
     reduce<U>(reducer: Reducer<T, U, [number]>, accumulator: U): U;
@@ -127,6 +137,7 @@ export namespace Collection {
     collect<U>(mapper: Mapper<T, Option<U>, [number]>): Indexed<U>;
     collectFirst<U>(mapper: Mapper<T, Option<U>, [number]>): Option<U>;
     some(predicate: Predicate<T, [number]>): boolean;
+    none(predicate: Predicate<T, [number]>): boolean;
     every(predicate: Predicate<T, [number]>): boolean;
     count(predicate: Predicate<T, [number]>): number;
     distinct(): Indexed<T>;
@@ -155,5 +166,19 @@ export namespace Collection {
     reverse(): Indexed<T>;
     join(separator: string): string;
     sortWith(comparer: Comparer<T>): Indexed<T>;
+    compareWith(iterable: Iterable<T>, comparer: Comparer<T>): Comparison;
+  }
+
+  export function sort<T extends Comparable<T>>(
+    collection: Indexed<T>
+  ): Indexed<T> {
+    return collection.sortWith(compareComparable);
+  }
+
+  export function compare<T extends Comparable<T>>(
+    a: Indexed<T>,
+    b: Iterable<T>
+  ): Comparison {
+    return a.compareWith(b, compareComparable);
   }
 }