+++ packages/alfa-list/src/list.ts
@@ -1,5 +1,6 @@
+import { Callback } from "@siteimprove/alfa-callback";
 import { Collection } from "@siteimprove/alfa-collection";
-import { Comparable, Comparer } from "@siteimprove/alfa-comparable";
+import { Comparable, Comparer, Comparison } from "@siteimprove/alfa-comparable";
 import { Hash, Hashable } from "@siteimprove/alfa-hash";
 import { Iterable } from "@siteimprove/alfa-iterable";
 import { Serializable } from "@siteimprove/alfa-json";
@@ -16,6 +17,7 @@ import * as json from "@siteimprove/alfa-json";
 import { Branch, Empty, Leaf, Node } from "./node";
 
 const { not } = Predicate;
+const { compareComparable } = Comparable;
 
 export class List<T> implements Collection.Indexed<T> {
   public static of<T>(...values: Array<T>): List<T> {
@@ -51,6 +53,10 @@ export class List<T> implements Collection.Indexed<T> {
 
   public isEmpty(): this is List<never> {
     return this._tail.isEmpty();
+  }
+
+  public forEach(callback: Callback<T, void, [number]>): void {
+    Iterable.forEach(this, callback);
   }
 
   public map<U>(mapper: Mapper<T, U, [number]>): List<U> {
@@ -123,6 +129,10 @@ export class List<T> implements Collection.Indexed<T> {
 
   public some(predicate: Predicate<T, [number]>): boolean {
     return Iterable.some(this, predicate);
+  }
+
+  public none(predicate: Predicate<T, [number]>): boolean {
+    return Iterable.none(this, predicate);
   }
 
   public every(predicate: Predicate<T, [number]>): boolean {
@@ -304,6 +314,10 @@ export class List<T> implements Collection.Indexed<T> {
 
   public sortWith(comparer: Comparer<T>): List<T> {
     return List.from(Iterable.sortWith(this, comparer));
+  }
+
+  public compareWith(iterable: Iterable<T>, comparer: Comparer<T>): Comparison {
+    return Iterable.compareWith(this, iterable, comparer);
   }
 
   public groupBy<K>(grouper: Mapper<T, K>): Map<K, List<T>> {
@@ -564,6 +578,13 @@ export namespace List {
   }
 
   export function sort<T extends Comparable<T>>(list: List<T>): List<T> {
-    return list.sortWith(Comparable.compare);
+    return list.sortWith(compareComparable);
+  }
+
+  export function compare<T extends Comparable<T>>(
+    a: List<T>,
+    b: Iterable<T>
+  ): Comparison {
+    return a.compareWith(b, compareComparable);
   }
 }