+++ packages/alfa-rules/src/sia-r15/rule.ts
@@ -6,6 +6,7 @@ import { Map } from "@siteimprove/alfa-map";
 import { Option } from "@siteimprove/alfa-option";
 import { Predicate } from "@siteimprove/alfa-predicate";
 import { Err, Ok } from "@siteimprove/alfa-result";
+import { Criterion } from "@siteimprove/alfa-wcag";
 import { Page } from "@siteimprove/alfa-web";
 
 import { expectation } from "../common/expectation";
@@ -22,6 +23,7 @@ const { and, not } = Predicate;
 
 export default Rule.Atomic.of<Page, Group<Element>, Question>({
   uri: "https://siteimprove.github.io/sanshikan/rules/sia-r15.html",
+  requirements: [Criterion.of("4.1.2")],
   evaluate({ device, document, response }) {
     return {
       applicability() {