+++ packages/alfa-rules/src/sia-r16/rule.ts
@@ -5,6 +5,7 @@ import { Element, Namespace } from "@siteimprove/alfa-dom";
 import { Iterable } from "@siteimprove/alfa-iterable";
 import { Predicate } from "@siteimprove/alfa-predicate";
 import { Ok, Err } from "@siteimprove/alfa-result";
+import { Criterion, Technique } from "@siteimprove/alfa-wcag";
 import { Page } from "@siteimprove/alfa-web";
 
 import { expectation } from "../common/expectation";
@@ -19,6 +20,7 @@ const { and, not, property } = Predicate;
 
 export default Rule.Atomic.of<Page, Element>({
   uri: "https://siteimprove.github.io/sanshikan/rules/sia-r16.html",
+  requirements: [Criterion.of("4.1.2"), Technique.of("ARIA5")],
   evaluate({ device, document }) {
     return {
       applicability() {