+++ packages/alfa-rules/src/sia-r47/rule.ts
@@ -4,6 +4,7 @@ import { Real } from "@siteimprove/alfa-math";
 import { None, Option } from "@siteimprove/alfa-option";
 import { Predicate } from "@siteimprove/alfa-predicate";
 import { Err, Ok } from "@siteimprove/alfa-result";
+import { Criterion } from "@siteimprove/alfa-wcag";
 import { Page } from "@siteimprove/alfa-web";
 
 import { expectation } from "../common/expectation";
@@ -15,6 +16,7 @@ const { and, equals } = Predicate;
 
 export default Rule.Atomic.of<Page, Element>({
   uri: "https://siteimprove.github.io/sanshikan/rules/sia-r47.html",
+  requirements: [Criterion.of("1.4.4"), Criterion.of("1.4.10")],
   evaluate({ document }) {
     return {
       applicability() {