+++ packages/alfa-rules/src/sia-r84/rule.ts
@@ -5,6 +5,7 @@ import { Predicate } from "@siteimprove/alfa-predicate";
 import { Refinement } from "@siteimprove/alfa-refinement";
 import { Ok, Err } from "@siteimprove/alfa-result";
 import { Style } from "@siteimprove/alfa-style";
+import { Criterion, Technique } from "@siteimprove/alfa-wcag";
 import { Page } from "@siteimprove/alfa-web";
 
 import { expectation } from "../common/expectation";
@@ -18,6 +19,11 @@ const { and } = Refinement;
 
 export default Rule.Atomic.of<Page, Element>({
   uri: "https://siteimprove.github.io/sanshikan/rules/sia-r84.html",
+  requirements: [
+    Criterion.of("2.1.1"),
+    Criterion.of("2.1.3"),
+    Technique.of("G202"),
+  ],
   evaluate({ device, document }) {
     return {
       applicability() {
@@ -62,36 +68,39 @@ export namespace Outcomes {
   );
 }
 
+/**
+ * Determine if an element is possibly scrollable. This is determined by the
+ * following factors:
+ *
+ * - A computed `width` or `height` that is not `auto`; and
+ * - A corresponding computed `overflow-x` or `overflow-y`, respectively, that
+ *   is `auto`, `clip`, or `scroll`.
+ */
 function isPossiblyScrollable(device: Device): Predicate<Element> {
-  /**
-   * Determine if an element is possibly scrollable. This is determined by the
-   * following factors:
-   *
-   * - A computed `width` or `height` that is not `auto`.
-   * - A computed `overflow-x` or `overflow-y` that is `auto`, `clip`, or
-   *   `scroll`.
-   */
+  const properties = [
+    ["overflow-x", "width"],
+    ["overflow-y", "height"],
+  ] as const;
+
   return (element) => {
     const style = Style.from(element, device);
 
-    if (
-      style.computed("width").value.type === "keyword" &&
-      style.computed("height").value.type === "keyword"
-    ) {
-      return false;
-    }
+    return properties.some(
+      ([overflow, dimension]) =>
+        style
+          .computed(dimension)
+          .some((dimension) => dimension.value !== "auto") &&
+        style.computed(overflow).some((overflow) => {
+          switch (overflow.value) {
+            case "auto":
+            case "clip":
+            case "scroll":
+              return true;
 
-    for (const property of ["overflow-x", "overflow-y"] as const) {
-      const { value: overflow } = style.computed(property);
-
-      switch (overflow.value) {
-        case "auto":
-        case "clip":
-        case "scroll":
-          return true;
-      }
-    }
-
-    return false;
+            default:
+              return false;
+          }
+        })
+    );
   };
 }