+++ packages/alfa-selector/test/selector.spec.tsx
@@ -4,8 +4,6 @@ import { test } from "@siteimprove/alfa-test";
 import { Selector } from "../src/selector";
 import { Context } from "../src/context";
 
-const context = Context.empty();
-
 test(".parse() parses a type selector", (t) => {
   t.deepEqual(Selector.parse("div").get().toJSON(), {
     type: "type",
@@ -821,10 +819,10 @@ test("#matches() checks if an element matches an :nth-child selector", (t) => {
     {d}
   </div>;
 
-  t.equal(selector.matches(a, context), true);
-  t.equal(selector.matches(b, context), false);
-  t.equal(selector.matches(c, context), true);
-  t.equal(selector.matches(d, context), false);
+  t.equal(selector.matches(a), true);
+  t.equal(selector.matches(b), false);
+  t.equal(selector.matches(c), true);
+  t.equal(selector.matches(d), false);
 });
 
 test("#matches() checks if an element matches an :nth-last-child selector", (t) => {
@@ -843,10 +841,10 @@ test("#matches() checks if an element matches an :nth-last-child selector", (t) 
     {d}
   </div>;
 
-  t.equal(selector.matches(a, context), false);
-  t.equal(selector.matches(b, context), true);
-  t.equal(selector.matches(c, context), false);
-  t.equal(selector.matches(d, context), true);
+  t.equal(selector.matches(a), false);
+  t.equal(selector.matches(b), true);
+  t.equal(selector.matches(c), false);
+  t.equal(selector.matches(d), true);
 });
 
 test("#matches() checks if an element matches a :first-child selector", (t) => {
@@ -861,8 +859,8 @@ test("#matches() checks if an element matches a :first-child selector", (t) => {
     {b}
   </div>;
 
-  t.equal(selector.matches(a, context), true);
-  t.equal(selector.matches(b, context), false);
+  t.equal(selector.matches(a), true);
+  t.equal(selector.matches(b), false);
 });
 
 test("#matches() checks if an element matches a :last-child selector", (t) => {
@@ -877,8 +875,8 @@ test("#matches() checks if an element matches a :last-child selector", (t) => {
     Hello
   </div>;
 
-  t.equal(selector.matches(a, context), false);
-  t.equal(selector.matches(b, context), true);
+  t.equal(selector.matches(a), false);
+  t.equal(selector.matches(b), true);
 });
 
 test("#matches() checks if an element matches an :only-child selector", (t) => {
@@ -898,8 +896,8 @@ test("#matches() checks if an element matches an :only-child selector", (t) => {
     Hello
   </div>;
 
-  t.equal(selector.matches(a, context), true);
-  t.equal(selector.matches(b, context), false);
+  t.equal(selector.matches(a), true);
+  t.equal(selector.matches(b), false);
 });
 
 test("#matches() checks if an element matches an :nth-of-type selector", (t) => {
@@ -920,10 +918,10 @@ test("#matches() checks if an element matches an :nth-of-type selector", (t) => 
     {d}
   </div>;
 
-  t.equal(selector.matches(a, context), true);
-  t.equal(selector.matches(b, context), false);
-  t.equal(selector.matches(c, context), true);
-  t.equal(selector.matches(d, context), false);
+  t.equal(selector.matches(a), true);
+  t.equal(selector.matches(b), false);
+  t.equal(selector.matches(c), true);
+  t.equal(selector.matches(d), false);
 });
 
 test("#matches() checks if an element matches an :nth-last-of-type selector", (t) => {
@@ -944,10 +942,10 @@ test("#matches() checks if an element matches an :nth-last-of-type selector", (t
     <span />
   </div>;
 
-  t.equal(selector.matches(a, context), false);
-  t.equal(selector.matches(b, context), true);
-  t.equal(selector.matches(c, context), false);
-  t.equal(selector.matches(d, context), true);
+  t.equal(selector.matches(a), false);
+  t.equal(selector.matches(b), true);
+  t.equal(selector.matches(c), false);
+  t.equal(selector.matches(d), true);
 });
 
 test("#matches() checks if an element matches a :first-of-type selector", (t) => {
@@ -963,8 +961,8 @@ test("#matches() checks if an element matches a :first-of-type selector", (t) =>
     {b}
   </div>;
 
-  t.equal(selector.matches(a, context), true);
-  t.equal(selector.matches(b, context), false);
+  t.equal(selector.matches(a), true);
+  t.equal(selector.matches(b), false);
 });
 
 test("#matches() checks if an element matches a :last-of-type selector", (t) => {
@@ -980,8 +978,8 @@ test("#matches() checks if an element matches a :last-of-type selector", (t) => 
     <div />
   </div>;
 
-  t.equal(selector.matches(a, context), false);
-  t.equal(selector.matches(b, context), true);
+  t.equal(selector.matches(a), false);
+  t.equal(selector.matches(b), true);
 });
 
 test("#matches() checks if an element matches a :only-of-type selector", (t) => {
@@ -1003,8 +1001,8 @@ test("#matches() checks if an element matches a :only-of-type selector", (t) => 
     <div />
   </div>;
 
-  t.equal(selector.matches(a, context), true);
-  t.equal(selector.matches(b, context), false);
+  t.equal(selector.matches(a), true);
+  t.equal(selector.matches(b), false);
 });
 
 test("#matches() checks if an element matches a :hover selector", (t) => {
@@ -1012,8 +1010,8 @@ test("#matches() checks if an element matches a :hover selector", (t) => {
 
   const p = <p />;
 
-  t.equal(selector.matches(p, context), false);
-  t.equal(selector.matches(p, context.hover(p)), true);
+  t.equal(selector.matches(p), false);
+  t.equal(selector.matches(p, Context.hover(p)), true);
 });
 
 test("#matches() checks if an element matches an :active selector", (t) => {
@@ -1021,8 +1019,8 @@ test("#matches() checks if an element matches an :active selector", (t) => {
 
   const p = <p />;
 
-  t.equal(selector.matches(p, context), false);
-  t.equal(selector.matches(p, context.active(p)), true);
+  t.equal(selector.matches(p), false);
+  t.equal(selector.matches(p, Context.active(p)), true);
 });
 
 test("#matches() checks if an element matches a :focus selector", (t) => {
@@ -1030,8 +1028,8 @@ test("#matches() checks if an element matches a :focus selector", (t) => {
 
   const p = <p />;
 
-  t.equal(selector.matches(p, context), false);
-  t.equal(selector.matches(p, context.focus(p)), true);
+  t.equal(selector.matches(p), false);
+  t.equal(selector.matches(p, Context.focus(p)), true);
 });
 
 test("#matches() checks if an element matches a :focus-within selector", (t) => {
@@ -1040,9 +1038,9 @@ test("#matches() checks if an element matches a :focus-within selector", (t) => 
   const button = <button />;
   const p = <p>{button}</p>;
 
-  t.equal(selector.matches(p, context), false);
-  t.equal(selector.matches(p, context.focus(p)), true);
-  t.equal(selector.matches(p, context.focus(button)), true);
+  t.equal(selector.matches(p), false);
+  t.equal(selector.matches(p, Context.focus(p)), true);
+  t.equal(selector.matches(p, Context.focus(button)), true);
 });
 
 test("#matches() checks if an element matches a :link selector", (t) => {
@@ -1054,11 +1052,11 @@ test("#matches() checks if an element matches a :link selector", (t) => {
     <area href="#" />,
     <link href="#" />,
   ]) {
-    t.equal(selector.matches(element, context), true, element.toString());
+    t.equal(selector.matches(element), true, element.toString());
 
     // Only non-visited links match :link
     t.equal(
-      selector.matches(element, context.visit(element)),
+      selector.matches(element, Context.visit(element)),
       false,
       element.toString()
     );
@@ -1066,7 +1064,7 @@ test("#matches() checks if an element matches a :link selector", (t) => {
 
   // These elements aren't links
   for (const element of [<a />, <p />]) {
-    t.equal(selector.matches(element, context), false, element.toString());
+    t.equal(selector.matches(element), false, element.toString());
   }
 });
 
@@ -1080,17 +1078,17 @@ test("#matches() checks if an element matches a :visited selector", (t) => {
     <link href="#" />,
   ]) {
     t.equal(
-      selector.matches(element, context.visit(element)),
+      selector.matches(element, Context.visit(element)),
       true,
       element.toString()
     );
 
     // Only visited links match :link
-    t.equal(selector.matches(element, context), false, element.toString());
+    t.equal(selector.matches(element), false, element.toString());
   }
 
   // These elements aren't links
   for (const element of [<a />, <p />]) {
-    t.equal(selector.matches(element, context), false, element.toString());
+    t.equal(selector.matches(element), false, element.toString());
   }
 });