+++ packages/alfa-sequence/src/nil.ts
@@ -1,4 +1,6 @@
+import { Comparer, Comparison } from "@siteimprove/alfa-comparable";
 import { Hash } from "@siteimprove/alfa-hash";
+import { Iterable } from "@siteimprove/alfa-iterable";
 import { Map } from "@siteimprove/alfa-map";
 import { None } from "@siteimprove/alfa-option";
 
@@ -15,6 +17,8 @@ export const Nil: Nil = new (class Nil {
   public isEmpty(): this is Sequence<never> {
     return true;
   }
+
+  public forEach(): void {}
 
   public map(): Nil {
     return this;
@@ -58,6 +62,10 @@ export const Nil: Nil = new (class Nil {
 
   public some(): boolean {
     return false;
+  }
+
+  public none(): boolean {
+    return true;
   }
 
   public every(): boolean {
@@ -174,6 +182,13 @@ export const Nil: Nil = new (class Nil {
 
   public sortWith(): Nil {
     return this;
+  }
+
+  public compareWith<T>(
+    iterable: Iterable<T>,
+    comparer: Comparer<T>
+  ): Comparison {
+    return Iterable.compareWith(this, iterable, comparer);
   }
 
   public equals(value: unknown): value is this {