+++ packages/alfa-sequence/src/sequence.ts
@@ -1,5 +1,6 @@
+import { Callback } from "@siteimprove/alfa-callback";
 import { Collection } from "@siteimprove/alfa-collection";
-import { Comparable, Comparer } from "@siteimprove/alfa-comparable";
+import { Comparable, Comparer, Comparison } from "@siteimprove/alfa-comparable";
 import { Lazy } from "@siteimprove/alfa-lazy";
 import { Map } from "@siteimprove/alfa-map";
 import { Mapper } from "@siteimprove/alfa-mapper";
@@ -11,10 +12,13 @@ import { Refinement } from "@siteimprove/alfa-refinement";
 import { Cons } from "./cons";
 import { Nil } from "./nil";
 
+const { compareComparable } = Comparable;
+
 export interface Sequence<T> extends Collection.Indexed<T> {
   // Collection<T> methods
 
   isEmpty(): this is Sequence<never>;
+  forEach(callback: Callback<T, void, [number]>): void;
   map<U>(mapper: Mapper<T, U, [number]>): Sequence<U>;
   flatMap<U>(mapper: Mapper<T, Sequence<U>, [number]>): Sequence<U>;
   reduce<U>(reducer: Reducer<T, U, [number]>, accumulator: U): U;
@@ -31,6 +35,7 @@ export interface Sequence<T> extends Collection.Indexed<T> {
   collect<U>(mapper: Mapper<T, Option<U>, [number]>): Sequence<U>;
   collectFirst<U>(mapper: Mapper<T, Option<U>, [number]>): Option<U>;
   some(predicate: Predicate<T, [number]>): boolean;
+  none(predicate: Predicate<T, [number]>): boolean;
   every(predicate: Predicate<T, [number]>): boolean;
   count(predicate: Predicate<T, [number]>): number;
   distinct(): Sequence<T>;
@@ -59,6 +64,7 @@ export interface Sequence<T> extends Collection.Indexed<T> {
   reverse(): Sequence<T>;
   join(separator: string): string;
   sortWith(comparer: Comparer<T>): Sequence<T>;
+  compareWith(iterable: Iterable<T>, comparer: Comparer<T>): Comparison;
 
   // Sequence<T> methods
 
@@ -138,6 +144,13 @@ export namespace Sequence {
   export function sort<T extends Comparable<T>>(
     sequence: Sequence<T>
   ): Sequence<T> {
-    return sequence.sortWith(Comparable.compare);
+    return sequence.sortWith(compareComparable);
+  }
+
+  export function compare<T extends Comparable<T>>(
+    a: Sequence<T>,
+    b: Iterable<T>
+  ): Comparison {
+    return a.compareWith(b, compareComparable);
   }
 }